requirements.txt
neo4j>=5.20.0
python-dotenv>=1.0.1
pandas>=2.2.2
opencc-python-reimplemented>=0.1.7
